+++ backend/homefinder/src/main/java/com/home/finder/homefinder/HomeFinderApplication.java
@@ -5,9 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class HomeFinderApplication {
-
 	public static void main(String[] args) {
 		SpringApplication.run(HomeFinderApplication.class, args);
 	}
-
 }
