+++ backend/homefinder/src/main/java/com/home/finder/homefinder/controller/UserController.java
@@ -2,7 +2,7 @@ package com.home.finder.homefinder.controller;
 
 import com.home.finder.homefinder.dto.UserDto;
 import com.home.finder.homefinder.entity.House;
-import com.home.finder.homefinder.service.UserService;
+import com.home.finder.homefinder.service.impl.UserServiceImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.validation.annotation.Validated;
@@ -13,10 +13,10 @@ import org.springframework.web.bind.annotation.*;
 @Validated
 public class UserController {
 
-    private UserService userService;
+    private UserServiceImpl userService;
     private static final Logger logger = LoggerFactory.getLogger(UserController.class);
 
-    public UserController(UserService userService) {
+    public UserController(UserServiceImpl userService) {
         this.userService = userService;
     }
