+++ backend/homefinder/src/main/java/com/home/finder/homefinder/entity/User.java
@@ -12,12 +12,9 @@ import org.springframework.security.core.userdetails.UserDetails;
 
 @Document(collection = "users")
 public class User implements UserDetails {
-
     @Transient
     public static final String SEQUENCE_NAME = "users_sequence";
-
     private List<House> favoriteHouses = new ArrayList<>();
-
     @Id
     private Long id;
     private String fullName;
@@ -60,5 +57,13 @@ public class User implements UserDetails {
     @Override
     public String getUsername() {
         return email;
+    }
+
+    public List<House> getFavoriteHouses() {
+        return favoriteHouses;
+    }
+
+    public void setFavoriteHouses(List<House> favoriteHouses) {
+        this.favoriteHouses = favoriteHouses;
     }
 }
