+++ backend/homefinder/src/main/java/com/home/finder/homefinder/service/SequenceGeneratorService.java
@@ -1,37 +1,5 @@
 package com.home.finder.homefinder.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.mongodb.core.MongoOperations;
-import org.springframework.stereotype.Service;
-
-import com.home.finder.homefinder.entity.DatabaseSequence;
-
-import static org.springframework.data.mongodb.core.FindAndModifyOptions.options;
-import static org.springframework.data.mongodb.core.query.Criteria.where;
-import static org.springframework.data.mongodb.core.query.Query.query;
-
-import org.springframework.data.mongodb.core.query.Update;
-
-import java.util.Objects;
-
-@Service
-public class SequenceGeneratorService {
-
-    private MongoOperations mongoOperations;
-
-    @Autowired
-    public SequenceGeneratorService(MongoOperations mongoOperations) {
-        this.mongoOperations = mongoOperations;
-    }
-
-
-    // In MongoDB, an upsert is a combination of update and insert: it updates a document if it exists, and inserts it if it doesn’t.
-
-    public long generateSequence(String seqName) {
-        DatabaseSequence counter = mongoOperations.findAndModify(query(where("_id").is(seqName)),
-                new Update().inc("seq",1), options().returnNew(true).upsert(true),
-                DatabaseSequence.class);
-        return !Objects.isNull(counter) ? counter.getSeq() : 1;
-    }
-    
+public interface SequenceGeneratorService {
+    Long generateSequence(String seqName);
 }
